# Give a chance to send a response before receiving next request on H2