ATS keeps reading H2 frames until either it reads all data in the receive buffer or it processes 128 H2 frames. This is not great if a client sends many requests at once. The first request will not be responded until ATS finishes/stops reading incoming frames, and that leads to a long TTFB. This also means ATS does not fully utilize a connection. ATS should be able to read following frames while response data is being transferred by Kernel/NIC.

This PR extends `Http2CommonSession::_should_do_something_else`, so ATS can stop reading frames regardless of the number of frames it read. I added `Http2CommonSession::interrupt_reading_frames()`, and by calling it after `Http2Stream::send_request`, ATS can get a chance to send response data before reading following H2 frames.


This benchmark was done on my laptop so rps and tps fluctuate and unreliable, but I can consistently see lower TTFB.

`h2load -n 3000 -m 100 -c 10 "https://localhost:8443/256k"`

Before
```
finished in 739.71ms, 4055.66 req/s, 1014.56MB/s
requests: 3000 total, 3000 started, 3000 done, 3000 succeeded, 0 failed, 0 errored, 0 timeout
status codes: 3000 2xx, 0 3xx, 0 4xx, 0 5xx
traffic: 750.47MB (786929450) total, 37.20KB (38090) headers (space savings 96.20%), 750.00MB (786432000) data
                     min         max         mean         sd        +/- sd
time for request:    34.20ms    364.11ms    192.23ms     75.67ms    72.47%
time for connect:    11.43ms     67.43ms     40.86ms     18.82ms    60.00%
time to 1st byte:   103.09ms    284.44ms    188.50ms     58.36ms    60.00%
req/s           :     405.98      458.36      432.78       16.96    60.00%
```

After
```
finished in 736.39ms, 4073.92 req/s, 1019.12MB/s
requests: 3000 total, 3000 started, 3000 done, 3000 succeeded, 0 failed, 0 errored, 0 timeout
status codes: 3000 2xx, 0 3xx, 0 4xx, 0 5xx
traffic: 750.47MB (786929450) total, 37.20KB (38090) headers (space savings 96.20%), 750.00MB (786432000) data
                     min         max         mean         sd        +/- sd
time for request:    14.43ms    371.40ms    200.59ms     66.48ms    75.73%
time for connect:     9.19ms     66.73ms     39.34ms     19.22ms    60.00%
time to 1st byte:    67.39ms    103.57ms     75.14ms     14.92ms    80.00%
req/s           :     407.46      517.14      458.63       36.60    60.00%
```

And this is from production. A service that makes a lot of concurrent requests. Download time per host within a same pop, and the first row is the host with this change.
<img width="865" alt="image" src="https://github.com/apache/trafficserver/assets/153144/fdd99784-16bb-4e71-99b6-02c5fb45f598">